refactor(v010): carve voicegw projects (listing) out of _legacy
Fifth per-command carve-out in section 2.

Adds voicegateway/cli/projects.py owning the 'projects' listing
command. The detail command 'project' joins the same module in the
next refactor commit rather than getting a separate file: both share
the project-related surface and natural cohesion outweighs the
strict one-command-per-file rule for this pair.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (7 source
files now).
